Allow multi-paragraph answers in accordions
Previously, the answers are inserted as a text node inside the accordion
panels. To improve semantics, we can treat answers as Markdown and
transform them to their corresponding HTML elements (so that a string of
text is converted into a `p` tag).

This also allows adding accordions with longer answers separated into
multiple paragraphs, which are neatly laid out thanks to the flow
utility.