Use aria-hidden to implement panel visibility
The toggle animation added in commit bb83fee sets the accordion panels
to `display: grid` which overrides the `hidden` attribute. This causes
the content of the collapsed panels to be accessible to screen readers,
thus introducing disparity between sighted and unsighted users.

Since we have implemented our own logic for toggling visibility, we can
switch to using `aria-hidden` for indicating state that is consistent
for sighted and unsighted users.